fix: break platform auth attribution import cycle
